+++ game/src/dlls/objects/223_Hagabon/Hagabon.c
@@ -157,7 +157,7 @@ void Hagabon_updateMovement(GameObject* obj, HagabonState* state) {
 
     (void)objMove(obj, obj->anim.velocityX * timeDelta, obj->anim.velocityY * timeDelta,
                   obj->anim.velocityZ * timeDelta);
-    (void)ObjAnim_AdvanceCurrentMove((int)obj, state->animSpeed, timeDelta, &animEvents.events);
+    (void)ObjAnim_AdvanceCurrentMove(obj, state->animSpeed, timeDelta, &animEvents.events);
 
     player = state->player;
     angle = (u16)getAngle(obj->anim.worldPosX - player->anim.worldPosX, obj->anim.worldPosZ - player->anim.worldPosZ);
