+++ game/src/dlls/objects/268_LanternFire/LanternFire.c
@@ -281,7 +281,7 @@ void LanternFireFly_update(GameObject* obj) {
     velocityPtr[2] = velocityPtr[2] * stepScale;
 
     if (LANTERN_FIREFLY_IS_FOLLOWING_PLAYER(state)) {
-        Sfx_KeepAliveLoopedObjectSound((int)obj, SFXTRIG_pk_lightcritter_lp);
+        Sfx_KeepAliveLoopedObjectSound(obj, SFXTRIG_pk_lightcritter_lp);
         if ((f32)state->timer > sLanternFireFlyEffectSpawnTimerThreshold) {
             if (state->stateId == LANTERN_FIREFLY_LIGHT_STATE_A || state->stateId == LANTERN_FIREFLY_LIGHT_STATE_B) {
                 (*gPartfxInterface)->spawnObject((void*)obj, LANTERN_FIREFLY_EFFECT_GLOW_A, NULL, 1, -1, NULL);
