+++ game/src/dlls/objects/390_CCgasventCo/CCgasventCo.c
@@ -96,7 +96,7 @@ u8 ccGasVentControl_countUnblockedVents(GameObject* obj, CCGasVentControlState* 
     }
     if (unblockedVentCount != 0) {
         if (state->loopedSoundActive == 0) {
-            Sfx_AddLoopedObjectSound((int)obj, SFXTRIG_en_diallp_c_223);
+            Sfx_AddLoopedObjectSound(obj, SFXTRIG_en_diallp_c_223);
             state->loopedSoundActive = 1;
         }
         Sfx_SetObjectSfxVolume(obj, SFXTRIG_en_diallp_c_223,
@@ -105,7 +105,7 @@ u8 ccGasVentControl_countUnblockedVents(GameObject* obj, CCGasVentControlState* 
                                CC_GAS_VENT_CONTROL_SFX_VOLUME_MAX);
     } else {
         if (state->loopedSoundActive != 0) {
-            Sfx_RemoveLoopedObjectSound((int)obj, SFXTRIG_en_diallp_c_223);
+            Sfx_RemoveLoopedObjectSound(obj, SFXTRIG_en_diallp_c_223);
             state->loopedSoundActive = 0;
         }
     }
